builtin/repack.c: remove redundant pack-based bitmaps
When we write a MIDX bitmap after repacking, it is possible that the
repository would be left in a state with both pack- and multi-pack
reachability bitmaps. This can occur, for instance, if a pack that was
kept (either by having a .keep file, or during a geometric repack in
which it is not rolled up) has a bitmap file, and the repack wrote a
multi-pack index and bitmap.

When loading a reachability bitmap for the repository, the multi-pack
one is always preferred, so the pack-based one is redundant. Let's
remove it unconditionally, even if '-d' isn't passed, since there is no
practical reason to keep both around.

Signed-off-by: Taylor Blau <me@ttaylorr.com>