builtin/repack.c: keep track of existing packs unconditionally
In order to be able to write a multi-pack index during repacking, 'git
repack' must keep track of which packs it wants to write into the MIDX.
This set is basically the existing packs which are not about to be
deleted, plus any new pack that was generated during the repack.

Prior to this patch, 'git repack' populated the list of existing packs
only when repacking all-into-one, but we will soon need to know this
list when repacking when writing a MIDX without a-i-o.

Populate the list of existing packs unconditionally, and guard removing
packs from that list only when repacking a-i-o.

Additionally, keep track of filenames of kept packs separately, since
this, too, will be used in an upcoming patch.

[For upstream, this patch should be split into the geometric parts and
the MIDX-repack parts, then squashed into 429f963fbb (builtin/repack.c:
add '--geometric' option, 2020-10-06) and cf6386cf1c (builtin/repack.c:
support writing a MIDX while repacking, 2020-10-07), respectively].