commit-graph.c: drop custom 'packed_oid_list' structure
To keep track of which commits are about to be written, the commit-graph
machinery first collects a list of packed, reachable, or explicitly
given object ids. It stores this data in a 'packed_oid_list'. This
data structure is unique to commit-graph.c, but provides identical
functionality to the standard oid_array.

There's no reason that commit-graph should maintain its own version of
the oid_array type, so for that reason, remove the definition of
packed_oid_list.

What follows is a mostly straightforward translation from one to the
other (e.g., replacing ALLOC_GROW + oidcpy() with oid_array_append() and
so on). There are two exceptions:

  - The pre-allocation in write_commit_graph() is dropped, since we can
    rely on 'oid_array_append()' to size the array appropriately, and it
    removes a potential spot where the two could diverge [1].

  - There's an extra copy in write_commit_graph() where we first copy a
    hash directly from a commit-graph into our own object_id, and then
    have oid_array_append() copies again. We could add a
    'oid_array_append_hash()' to avoid the secondary copy, but it
    doesn't seem worth it (we're only adding one additional hashcpy()
    per commit in the most recent commit-graph layer).

[1]: While testing this patch and before I removed the pre-allocation
logic, I found a bug where we allowed '.alloc' to drift, and then we
segfaulted within oid_array_append().

Co-authored-by: Jeff King <peff@peff.net>