commit-graph.c: drop 'packed_commit_list' structure
This should be able to be an oid_array as well, since we can easily
re-lookup the commits and they'll come back to us quickly in an already
parsed state.

Two nitches here:

  - First, this just doesn't work. There's some bug, but I don't have
    the time to figure out what it is.

  - Second, and more pressing, is that there doesn't seem to be a need
    to have this at all; we ought to be able to just filter the original
    oid array and use that.

Anyway, I'm leaving this on a -wip branch for now so that it doesn't get
integrated into my 'private' branch in the meantime.