blame-tree: document blame-tree-cache format
Create a chunk-format file for the blame-tree cache, and place those
files in the $GIT_DIR/objects/info/blame-tree/ directory. This allows
sharing the caches across fork boundaries.

The header information helps differentiate this file from others if
separated from its filename and location. Version information helps it
stay robust to future changes.

The chunks can be read at different times: we can quickly check the
COMMIT chunk to see if it is the all-zeroes OID and skip reading the
rest of the file on read. If we are updating all cached files, then we
only need to read the METADATA chunk in order to generate a new command
to recompute the file. We can delay reading the RESULTS chunk until we
have hit our cached commit during our walk.