ci/install-dependencies.sh: install coccinelle from universe
The apt package "coccinelle" moved to the universe repository in the
package mirrors that are used by GitHub Actions. This causes
coccinelle to be reported as missing, since it is no longer available in
the main repository.

Since universe isn't sourced by default, add it as an apt repository
before updating and installing coccinelle.

Signed-off-by: Taylor Blau <me@ttaylorr.com>