builtin/multi-pack-index.c: support --stdin-packs mode
A future caller will want to write a multi-pack index containing only
some of the packs in the object directory. This can occur for a variety
of reasons, but the specific caller in mind when writing this commit is
'git repack'.

'git repack' will want to write a multi-pack index containing a certain
set of packs in the repository. Of course, this multi-pack index will be
written after generating any new packs, but it should come before any
existing packs may be deleted (in the case of '-d').

This is because it is possible that some of the packs about to be
deleted may also be included in an existing multi-pack index. If this is
the case, then the existing MIDX becomes invalid. Instead, make sure
that the new MIDX is written first, so that there is no point during
which the repository doesn't have a MIDX.

This requires that the MIDX being written contain only some of the packs
found in the object directory. Namely, it should contain anything that
*isn't* about to be deleted.

For this purpose, add a '--stdin-packs' mode, which allows the caller
to specify which packs they wish to include in a multi-pack index.

Note that the method implemented and exposed by the previous commit
could very well be called directly by 'git repack', but this introduces
some awkward work-arounds related to having opened and closed the object
store. Generating a multi-pack index in a sub-process allows 'git
multi-pack-index' to get a fresh look at the current state of the object
store, which avoids many of these issues.

Signed-off-by: Taylor Blau <me@ttaylorr.com>