midx: preliminary support for '--refs-snapshot'
To figure out which commits we can write a bitmap for, the multi-pack
index/bitmap code does a ref traversal, and marks any commits that it
finds in its own list of object as eligible to receive a bitmap.

This approach will cause a problem when multi-pack bitmaps are able to
be generated from 'git repack'. Here, it's possible that the ref tips
have moved on from where they were when the repack started. Even though
we ignore commits that don't exist in the MIDX (when doing a scan of the
ref tips), it's possible that a commit in the MIDX reaches something
that isn't.

This can happen when a multi-pack index contains some pack which refers
to loose objects (which by definition aren't included in the multi-pack
index).

By taking a snapshot of the references before we start repacking, we can
close that race window. In the above scenario (where we have a packed
object pointing at a loose one), we'll either (a) take a snapshot of the
references before seeing the packed one, or (b) take it after, at which
point we can guarantee that the loose object will be packed and included
in the MIDX.

This patch does just that. It writes a temporary "reference snapshot",
which is a list of OIDs that are at the ref tips before writing a
multi-pack bitmap. References that are "preferred" (i.e,. are a suffix
of at least one value of the 'pack.preferBitmapTips' configuration) are
marked with a special '+'.

The format is simple (and yet undocumented): one line per commit at each
tip, with an optional '+' at the beginning (for preferred references, as
described above).