blame-tree: create blame_tree_cache_writer
This struct will be used to store the results in-memory and in a later
change write the results to the file format. For now, when the
BLAME_TREE_CACHE flag is specified, then initialize it and redirect the
output to filling its array of results.

While we are initializing this struct, special case it to only care when
there are is at most one pathspec. (The full tree is the "zero" pathspec
case.)