blame-tree: update option parsing
Use the parse-opt library to pull some blame-tree specific options. The
only one we are adding right now is the --cache option. This does
nothing at the moment except redirect the output into an array. We will
update this option to actually write a file at the end of the process.