blame-tree: auto-create placeholder cache files
We automatically cache the blame-tree results for the root directory
because that is the most common case, and is most likely to be slow.

However, some repositories have slow cases that are not the root
directory. We cannot predict these except by real-time performance
characteristics.

Add a new capability to the blame-tree code to detect when a non-
empty pathspec is uncached, and use a timer to detect if we go
beyond the limit during our commit walk. If we do, then write a
placeholder cache file.

Since we do not have a real commit object for our placeholder, use the
null_oid instead when writer->commit is NULL.

Add a new blameTree.limitMilliseconds config option that specifies
how long to wait before writing these cache files. This can be
adjusted to have a shorter or longer timeout than the default of
1000ms (1 second).

We test the timeout during the commit walk to ensure that we still
write these files even if the process is terminated due to a
timeout.

When we write the placeholder, set the new goal_end_time to zero
to prevent future checks to rewrite the placeholder.