blame-tree: limit cache count for --update-cache
If we have very large repositories with many directories that need
caching, then don't update them all on every maintenance. Let some of
the cache files get stale and limit ourselves to 10 updates per run.
This can be adjusted by the new config option blameTree.maxWrites.

List the cacehd directories and sort them by mtime. Update the oldest
cache files to ensure we rotate the files appropriately. This ordering
is hard to test in the test suite without inserting sleep()s, so instead
the number of subcommands is verified.

The root directory is special: always update it. Since it appears in the
file list, we need to not count it when iterating over our loop. Hence
the use of write_count and the change to update_cache() to set a
pass-by-reference to signal whether the pathspec was nonempty.

To test that this works, use 'test-tool chmtime'. In order to prevent
issues with the mtime getting mixed up, ensure that the mtimes are split
by at least two seconds each, and have order according to the loop.