blame-tree: read from cache during write
I noticed that some of the blame-tree cache commands are quite slow.
These are likely very extreme data shapes, but they will just keep being
slow until we let the 'git blame-tree --cache' command read from the
existing cache to speed up its execution.

The first implementation of the blame-tree cache didn't allow this for
two reasons: it didn't use a lockfile, and it wrote the results to the
file during the commit walk.

The new implementation _does_ use a lockfile, but in addition it
populates an in-memory list of the results to write at the end of the
process. Inside blame_tree_release(), the reader is freed before the
writer does its write, which ensures that this would work on Windows (if
we ever care about that in the future).

Add a new config option, blameTree.skipReadCache, that allows us to
disable reading from the existing cache files. This serves as a safety
valve if we ever need to disable the cache in production quickly.