builtin/pack-objects.c: support creating cruft packs
In the next patch, `git repack` will gain a new `--cruft` option,
allowing it to pack unreachable objects separately. But the majority of
the work to create a cruft pack will be done by `pack-objects` using
what is implemented in this patch.

`git pack-objects --cruft` creates a new pack (and an `.mtimes` file)
containing objects that aren't found in any `--keep-pack`. To create a
cruft pack `git repack` will pass the names of any packs which are not
getting deleted as `--keep-pack`s. Since `git repack` usually only
creates packs which contain reachable objects, objects that aren't
contained in those packs are known to be unreachable.

Accumulating cruft objects works roughly as follows:

  - Add every non-expired loose and packed object not contained in a
    kept pack to a traversal.

  - Run that traversal, adding objects that we encounter along the way
    to the new cruft pack, and stopping once we hit an object contained
    in any kept pack(s).

The timestamp of an object is determined using the following rules:

  1. If an object is found loose, then its timestamp is the mtime of the
     loose object file.

  2. If an object is found in a cruft pack, then its timestamp can be
     read out of the corresponding `.mtimes` file.

  3. Otherwise, the object belongs to a non-cruft pack, in which case
     its timestamp is the mtime of that pack.

Why is the reachability traversal necessary? It is done in order to
"rescue" cruft objects which would have otherwise been expired (when
given a finite `--cruft-expiration`). Consider an example where a commit
(and the trees and blobs reachable from it) is about to be expired, but
the commit is freshened. In this case we want to avoid pruning the trees
and blobs even though they may have mtimes occurring earlier than the
grace period, since there is another object from which they are
reachable which we are not expiring. The traversal provides the
reachability information necessary to discover which would-be
expire-able objects are rescued by non-expired ones.

Confusingly, objects do not propagate their mtimes to objects that they
can reach. Doing so is tricky, but luckily for us, it is also
unnecessary, since non-expired objects always rescue any reachable ones.
That's because we include all objects visited along a traversal
(including expired ones, which are only discarded when considering
whether or not to add a tip to the traversal).

Finally, stopping early once an object is found in a kept pack is safe
to do because the kept packs ordinarily represent which packs will
survive after repacking. Assuming that it _isn't_ safe to halt a
traversal early would mean that there is some ancestor object which is
missing, which implies repository corruption (i.e., the complete set of
reachable objects isn't present).

(cherry picked from commit 2f14c0b523f166d1b89d3648a0aea7ebe54152e3)