p5310: write an on-disk reverse index
Git must generate an on-disk reverse index for the pack corresponding to
a .bitmap before it can use that bitmap.

In repositories that store many objects in the bitmapped pack, this can
a meaningful amount of time. It doesn't seem to make much of a
measurable difference when setting GIT_PERF_LARGE_REPO to a local copy
of the kernel, but it does make a big difference when using a repo
containing the kernel's objects, and the objects from all forks of the
kernel.

    [todo; performance numbers here]

Alternatively, this patch could be dropped and replaced instead with one
which starts writing on-disk reverse indexes everywhere by default. But
a more minimal approach is taken here.