p5310: add a test to load a single commit
Measure the time it takes to produce a rev-list answer from a single
commit, with the hope that that commit has a bitmap (or is near a
bitmapped commit).

(In practice this assumption is reasonable, since we try to place
bitmaps near the ref tips as much as possible. It's possible that HEAD
wasn't bitmapped, and we could be more precise by using the 'bitmap'
test-tool, but it seems like overkill for what we're trying to do).

This test should give us a better understanding of what the overhead of
working with bitmaps is. Since we expect to do no or very little
follow-on traversal between the traversal tips and bitmapped commits,
this should measure about how much time it takes to load a bitmap.