reachable.c: ignore objects in kept packs, too
With the recent fix in 73519fc427 (for_each_packed_object: add missing
check for in-core kept packs, 2021-07-28), that function is clever
enough to avoid looking at in-core kept packs when we ask to skip them.

But in the case of add_unseen_recent_objects_to_traversal(), we do not
ignore the copies of objects found in non-kept packs when copies of that
same object also exist in a kept pack.

This is an important optimization for cruft packs, since all reachable
objects will be found both in the kept and non-kept packs.

Suggested-by: Jeff King <peff@peff.net>
(cherry picked from commit f91a35a1043cd2c85427b869a83434d109729a20)