t5327: demonstrate MIDX bitmaps and cruft objects race
When using cruft packs, the following race can occur when a geometric
repack that writes a MIDX bitmap takes place afterwords:

  - First, create an unreachable object and do an all-into-one cruft
    repack which stores that object in the repository's cruft pack.
  - Then make that object reachable.
  - Finally, do a geometric repack and write a MIDX bitmap.

Assuming that we are sufficiently unlucky as to select a commit from the
MIDX which reaches that object for bitmapping, then the `git
multi-pack-index` process will complain that that object is missing.

The reason is because we don't include cruft packs in the MIDX when
doing a geometric repack. Since the "make that object reachable" doesn't
necessarily mean that we'll create a new copy of that object in one of
the packs that will get rolled up as part of a geometric repack, it's
possible that the MIDX won't see any copies of that now-reachable
object.

Of course, it's desirable to avoid including cruft packs in the MIDX
because it causes the MIDX to store a bunch of objects which are likely
to get thrown away. But excluding that pack does open us up to the above
race.

This is particularly common in our infrastructure because the resulting
trees generated by libgit2's test merges can be included in the cruft
pack if a cruft repack takes place before the PR for which the
test-merge was generated hasn't yet been merged. When the PR is merged,
libgit2 freshens the cruft pack instead of writing it loose, because it
doesn't understand cruft packs.

This patch demonstrates the bug. The subsequent patch will remove the
test_must_fail.

(cherry picked from commit e2edec4ec0b5d952aa45fb750303a6c1150f64b4)