builtin/pack-objects.c: tolerate missing cruft objects
Since Git only expects validity for reachable objects, it's perfectly
reasonable for unreachable objects to be missing/corrupt/etc. So
generating a cruft pack should just unreachable objects for which that's
the case, but it currently vomits instead.

That's because of the reachability traversal we do over the unreachable
objects to rescue objects which would have otherwise been deleted. That
reachability traversal should ignore missing/corrupt objects, so pass
the magic 'ignore_missing_links' bit to get it to do that.

That does mean that if we have 'A -> B -> C' (meaning that 'C' is
reachable from 'B' and so on) with 'C' about to be pruned, 'A' able to
rescue it and 'B' missing/corrupt, then 'C' won't get rescued. But
without 'B' there, we pack-objects can't discover the transitive link
between 'A' and 'C', so there is nothing we can do to rescue 'C' in that
case.

Of course, if those objects became reachable again, then the corruption
would escalate to the level of repo-corruption, but cruft packs are only
concerned with objects in their unreachable state, so it's not a problem
for us here.

(cherry picked from commit 1b9fff0a7570dca091c4e72ff741b0099da7d8b3)