builtin/pack-objects.c: avoid cruft object traversal without expiration
When generating a cruft pack with an expiration (in other words, when
there is a possibility that we might remove objects) we do a
reachability traversal using all cruft objects as tips in order to
rescue unreachable but reachable-from-recent objects from deletion.

But when we have no cruft expiration, no such traversal is needed: we
will already generate the correct list of objects to save in the cruft
pack. Since we won't delete objects without an expiration, there is no
need to have objects rescue others, and no need for a reachability
traversal.

Implement this lightweight mode of generating cruft objects by building
on the same functions used to implement `--keep-unreachable`:
`add_objects_in_unpacked_packs()` and `add_unreachable_loose_objects()`.
The only changes necessary for these functions is to compute and store
the mtimes when in cruft-pack mode.

For an example, consider a copy of git.git with all of linux.git's
objects fetched into it (so that there are many objects which don't have
any references pointing at them).

  - Repacking that repository all-into-one with `git repack -adnk` takes
    49.385 seconds, 15.634 of which are spent in enumeration, 2.827 in
    counting, 3.401 in compression, and 26.935 in writing.

  - Repacking again (using the tip of `github`) with `git repack --cruft
    -d` takes a whopping 97.66 seconds, with the time broken down
    between generating the main and cruft packs as follows:

                | main  | cruft
    ------------+-------+--------
    enumerating | 2.599 | 22.942
       counting | 0.349 |  2.570
      traversal |   -   | 38.996
    compression | 0.067 |  1.166
        writing | 1.283 | 26.505

  - Versus repacking with this commit using the same command (starting
    from the state generated by `git repack -adnk`) takes 46.386 seconds
    (around 3 seconds *less* than the classic all-into-one repack) and
    spends its time like so:

                | main  | cruft
    ------------+-------+--------
    enumerating | 2.581 | 10.605
       counting | 0.346 |  2.602
      traversal |   -   |      0
    compression | 0.065 |  1.167
        writing | 1.262 | 27.163

    The totals in each region are comparable to the all-into-one repack,
    but we spend a little less time enumerating and compressing overall,
    leading to a modest win over the existing repacking strategy.

There's a little more code duplication than I'd like (in particular, I'm
most uncomfortable with repeating the logic to determine the mtime of a
packed object. But I think it's not too bad overall, and we do share a
lot of code with the existing functions used to implement
`--keep-unreachable`, which is good.

Note that a side-effect of avoiding the traversal is that we will not
generate name-hashes for deltas of cruft objects. But that's an OK
trade-off, since that's exactly what `git repack -k` does, and what we
have been relying on and serving for years.

(cherry picked from commit 4536664b39215435e3c703e51ddee54cd541d496)