builtin/pack-objects.c: tolerate missing blobs in cruft packs
When we encounter a cruft tree during the reachability traversal we
perform over cruft objects, we rescue all of that tree's entries, adding
objects we haven't seen before as a side-effect.

This includes objects which don't exist in the repository. If the
missing object is a tree, then we'll be OK, since we set
ignore_missing_links in the traversal (so list-objects will silently
ignore them). But if we have a missing blob, then we'll try to pack an
object that we don't have without noticing it earlier.

We can handle this case by catching it just before we add the object to
our packing list in add_cruft_object_entry(). Note that we only have to
look for an object if we:

  - didn't just find it in a pack via want_object_in_pack(), since
    finding it in a pack necessarily means that we have a good copy of
    the object

  - know the object type is a blob, since we'll handle missing sub-trees
    earlier in the traversal code

  - can't find the object loose, meaning that (since if we got to this
    point, we know we also don't have a packed copy) the object is truly
    missing.

Because of our recent changes, this only occurs when we do a
reachability traversal over the cruft objects (i.e., when dropping
objects with `--cruft-expiration`).

(cherry picked from commit e88f685e676a2002bd013192b3566b8f37e45f97)