builtin/repack.c: handle object corruption among reachable objects
When repacking with `--cruft`, it is possible that the pack containing
reachable objects couldn't be generated, for example because another
pack containing some reachable objects is corrupt.

If this is the case, the first pack-objects does not generate anything,
leaving `names` empty. This is different from the behavior when
generating a pack with just `-Ad`, since unreachable objects are
loosened, at which point `pack-objects` discovers the corruption,
figures out that it cannot loosen the object, and fails with a non-zero
exit code.

But none of that happens in the cruft pack case, since we deliberately
don't force the original `pack-objects` to loosen the existing unreachable
objects. Here, `pack-objects` exits cleanly, having generated nothing,
and then we short circuit writing a cruft pack because our conditional
is:

    if ((pack_everything & PACK_CRUFT) && names.nr)

Then we proceed to delete all of the packs, including non-corrupt ones.
Note that this bug, though serious, is exceedingly rare, since it
requires that there are no other new objects (because it relies on not
generating any packs).

But, we can remedy this by trying to generate a cruft pack anyway, since
doing so will cause us to notice the corruption in the same way that the
"force unreachable objects loose" option of `git repack -Ad` works, and
fail the repack before we have the opportunity to delete any packs. In
other words, we try to move any cruft objects into the new pack, notice
that we can't, and then fail the operation.

To trigger this bug, we need to play a few tricks in order to make sure
that corrupt objects aren't noticed too early (e.g., by existing in the
index, reflog, or at the ref tips).

(cherry picked from commit db0a27f9c09b563c958889296d5eda4a8bd0da4c)