blame-tree: provide pathspec during cache write error
In a completely unrelated investigation, running maintenance in
production revealed the following error:

  fatal: max-depth cannot be used with wildcard pathspecs
  warning: error while computing cache file '0a1ec53d0c58166479a404cd3f555fc81a33e9b0.btc'

Looking into this issue, I was able to open up the offending file and
see that it was a placeholder cache file (the commit chunk had the null
OID, signalling that it was a request to start cachine, not a
successfully written cache file full of data). I copied that .btc file
to my local repository and ran 'git blame-tree --update-cache HEAD' and
it successfully updated the .btc file. It still had no results because
that path did not exist at HEAD, but the commit chunk was updated to
include HEAD.

The filename loaded by this file doesn't have anything strange about it
that would make this pathspec look strange, and it did not show that
warning on my machine. I was expecting that if the file itself was
problematic that I would get that warning.

This change only adds a new warning deeper in the stack so we can access
the exact pathspec string being used, which could help identify the
problem.